+++ alpha/config/kConfMapNames.php
@@ -14,4 +14,5 @@ class kConfMapNames
 	const ANALYTICS = 'analytics';
 	const RUNTIME_CONFIG = 'runtime_config';
 	const DYNAMIC_EMAIL_CONTENTS = 'dynamic_email_contents';
+	const REDIS = 'redis';
 }