+++ plugins/game/lib/api/filters/base/KalturaUserScorePropertiesBaseFilter.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @package plugins.game
+ * @relatedService UserScoreService
+ * @subpackage api.filters.base
+ * @abstract
+ */
+abstract class KalturaUserScorePropertiesBaseFilter extends KalturaRelatedFilter
+{
+}